removed circular dependencies and madgerc override
