+++ src/core/config/remote/remote_config_data.ts
@@ -1,0 +1,29 @@
+/**
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the ""License"");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an ""AS IS"" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+import type {RemoteConfigMetadata} from './metadata.js';
+import type {ComponentsDataWrapper} from './components_data_wrapper.js';
+import type {CommonFlagsDataWrapper} from './common_flags_data_wrapper.js';
+import {type Cluster, type Namespace} from './types.js';
+
+export interface RemoteConfigData {
+  metadata: RemoteConfigMetadata;
+  clusters: Record<Cluster, Namespace>;
+  components: ComponentsDataWrapper;
+  lastExecutedCommand: string;
+  commandHistory: string[];
+  flags: CommonFlagsDataWrapper;
+}
